temporarily Split development and complete report details page
