+++ src/components/testreports/spocktestsindevelopment/DevelopmentReportDetails.js
@@ -4,35 +4,22 @@ import { firestoreConnect } from 'react-redux-firebase'
 import { compose } from 'redux'
 import moment from 'moment'
 import { Redirect } from 'react-router-dom'
-import { downloadReport } from "../../store/actions/reportActions";
+import {downloadReport} from "../../../store/actions/reportActions";
 
-import './reportdetails.css';
+import '../../testreports/reportdetails.css';
 
 /**
- * NOTE: Couldnt get the correct collectionUrl to use with firestoreConnect. Created two different classes instead:
- * CompleteReportDetails and DevelopmentReportDetails. Will revisit this later!
+ * NOTE: Refactor so we can only use one ReportDetails class
  */
 
-let collectionUrl = getCollectionUrl();
-const ReportDetails = (props) => {
-  // if((window.location.href).includes('completed') ){
-  //   collectionUrl = 'completedreports'
-  // } else if((window.location.href).includes('development') ){
-  //   collectionUrl = 'developmentreports'
-  // }
-  const { collectionUrlProp, auth, report, downloadReport, reportDownload} = props;
+const DevelopmentReportDetails = (props) => {
+  const {auth, report, downloadReport, reportDownload} = props;
   if (!auth.uid) return <Redirect to='/login' />;
-
-  collectionUrl = collectionUrlProp;
-
-  console.log('in ReportDetails');
-  console.log(collectionUrl);
 
   if (report) {
     console.log("Report Sent");
     console.log(report);
     downloadReport(report);
-    // downloadReport({url: 'url'});
 
     var htmlDoc = {__html: reportDownload};
 
@@ -65,9 +52,7 @@ const mapStateToProps = (state, ownProps) => {
   console.log('state in report details');
   console.log(state);
   const id = ownProps.match.params.id;
-  const reports =  getCollectionUrl() == 'completedreports'
-      ?state.firestore.data.completedreports
-      :state.firestore.data.developmentreports;
+  const reports =  state.firestore.data.developmentreports;
 
   const report = reports ? reports[id] : null;
 
@@ -89,22 +74,14 @@ const mapDispatchToProps = (dispatch) => {
   }
 };
 
-function getCollectionUrl(){
-  if((window.location.href).includes('completed') ){
-    return 'completedreports'
-  } else if((window.location.href).includes('development') ){
-    return 'developmentreports'
-  }
-}
-
 export default compose(
     connect(mapStateToProps, mapDispatchToProps),
     firestoreConnect([
       {
         collection: 'company',
         doc: 'tala',
-        subcollections: [{ collection: collectionUrl}],
-        storeAs: collectionUrl
+        subcollections: [{ collection: 'developmentreports'}],
+        storeAs: 'developmentreports'
       }
     ])
-)(ReportDetails)
+)(DevelopmentReportDetails)
